src/containers/steering_gear_system/requirements.txt
pymodbus[serial]==3.0.0.dev4
twisted
nats-py==2.2